rm double and fix tests
